refactor: migrate Map into ts
